feat: add group member attribute
Hi, I saw your nice simple little library for LDAP authentication and started to play around with it. Nice work! I was particularly interested in extracting group information and saw that your example for the forumsys LDAP server wasn't working properly. I looked at the code and saw that the search you do is based on the object class and member attributes, but this last attribute should be uniqueMember for forumsys. So I thought of adding a groupMemberAttribute option. Then, when passing the groupClass option as groupOfUniqueNames and groupMemberAttribute as uniqueMember it finds the group information correctly. Sorry if you find this an unnecessary change, but at least it will make it clear to other users why it won't find group information based on your example? Anyways, glad to hear your thoughts and have a nice evening.